requirements.txt
# Server
fastapi==0.115.6
uvicorn==0.34.0

# Environment
python-dotenv==1.0.1

# Data Processing
pandas>=2.2.3
numpy>=2.1.0
openpyxl==3.1.5

# NLP & Machine Learning
transformers>=4.46.0
torch>=2.5.1
scipy>=1.14.1

# File Upload
python-multipart==0.0.20
